+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3extralearners dev
 
-* Add `formula` and `anc` params to `surv.flexible` learner, as well as `response` predict type (mean survival time)
+* Add `formula` and `anc` params to `surv.flexible` learner, as well as `response` predict type (mean survival time).
+* Fix `regr.gamboost` regression predictions (#498).
 
 # mlr3extralearners 1.2.0